requirements.txt
appdirs==1.4.4
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
importlib-metadata==6.7.0
lxml==4.9.2
Markdown==3.4.3
multitasking==0.0.11
numpy==1.24.3
pandas==2.0.2
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
webencodings==0.5.1
yfinance==0.2.20
zipp==3.15.0
